fixed minor bugs in time statements and variable save in wq3d

+++ Src/VELPLTH.f
@@ -296,7 +296,7 @@ C
           JSVPH=0  
         ENDIF  
         IF(ISDYNSTP.EQ.0)THEN  
-          TIME=DT*FLOAT(N)+TCON*TBEGIN  
+          TIME=INT(DT*FLOAT(N)+TCON*TBEGIN)  
         ELSE  
           TIME=TIMESEC  
         ENDIF
